Rename comments api file
